fix: fix/export-hugo-static
